Ginthi_Backend-development/ginthi_agents/lambda_functions/web_scraper/vendor_master/requirements.txt
playwright==1.45.0
boto3
python-dotenv
pandas
botocore